Add base tab panel
