+++ src/renderer/components/animator/OverlayTabPanes/OverlayTabPanePlayback/OverlayTabPanePlayback.css
